Update planet creation and addition process
The code has been revised to facilitate the creation of a new planet and its addition to the 'universe'. The new planet can now be created in the NewPlanetStage and added directly to the universe object. The ability to quit the program has also been improved, and certain key functionalities have been reassigned for user convenience.