Update to specific Python version and fix typo
The Python version has been specified to Python 3.12 within the Solar System project for consistency and compatibility. Fixed a typo with apostrophe use in degrees comment in Universe.py. Also, the pygame version has been specified in the requirements.txt file for reliability, and the README file has been created with project setup instructions and a brief description of the project.